+++ src/main/java/com/xxxmkxxx/customgui/client/ui/controls/button/SimpleButton.java
@@ -92,17 +92,17 @@ public class SimpleButton extends AbstractButton implements LeftClickEventHandle
 
     @SuppressWarnings("DuplicatedCode")
     private void updateIndents() {
-        int leftButtonMargin = style.getMargins().getLeft();
-        int topButtonMargin = style.getMargins().getTop();
-        int leftButtonPadding = style.getPaddings().getLeft();
-        int topButtonPadding = style.getPaddings().getTop();
-        int leftTextMargin = text.getStyle().getMargins().getLeft();
-        int topTextMargin = text.getStyle().getMargins().getTop();
+        float leftButtonMargin = style.getMargins().getLeft();
+        float topButtonMargin = style.getMargins().getTop();
+        float leftButtonPadding = style.getPaddings().getLeft();
+        float topButtonPadding = style.getPaddings().getTop();
+        float leftTextMargin = text.getStyle().getMargins().getLeft();
+        float topTextMargin = text.getStyle().getMargins().getTop();
 
         frame.moveStartPos(leftButtonMargin, topButtonMargin);
 
-        int tempXDistance = leftButtonMargin + leftButtonPadding + leftTextMargin;
-        int tempYDistance = topButtonMargin + topButtonPadding + topTextMargin;
+        float tempXDistance = leftButtonMargin + leftButtonPadding + leftTextMargin;
+        float tempYDistance = topButtonMargin + topButtonPadding + topTextMargin;
 
         text.getFrame().moveStartPos(tempXDistance, tempYDistance);
         text.getFrame().moveStopPos(tempXDistance, tempYDistance);
